weakreference found should remove subscription
